// api/prisma.d.ts
 declare const prisma: unique symbol;
 export type PrismaPromise<A> = Promise<A> & { [prisma]: true };
 type UnwrapPromise<P extends any> = P extends Promise<infer R> ? R : P;
 type UnwrapTuple<Tuple extends readonly unknown[]> = {
   [K in keyof Tuple]: K extends `${number}`
     ? Tuple[K] extends PrismaPromise<infer X>
       ? X
       : UnwrapPromise<Tuple[K]>
     : UnwrapPromise<Tuple[K]>;
 };
 
 /**
  * Model User
  */
 
 export type User = {
   id: number;
   email: string;
   name: string | null;
   role: Roles;
 };
 
 /**
  * Enums
  */
 
 // Based on
 // https://github.com/microsoft/TypeScript/issues/3192#issuecomment-261720275
 
 export const Roles: {
   admin: 'admin';
   user: 'user';
 };
 
 export type Roles = typeof Roles[keyof typeof Roles];
 
 /**
  * ##  Prisma Client ʲˢ
  *
  * Type-safe database client for TypeScript & Node.js
  * @example
  * ```
  * const prisma = new PrismaClient()
  * // Fetch zero or more Users
  * const users = await prisma.user.findMany()
  * ```
  *
  *
  * Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client).
  */
 export class PrismaClient<
   T extends Prisma.PrismaClientOptions = Prisma.PrismaClientOptions,
   U = 'log' extends keyof T
     ? T['log'] extends Array<Prisma.LogLevel | Prisma.LogDefinition>
       ? Prisma.GetEvents<T['log']>
       : never
     : never,
   GlobalReject = 'rejectOnNotFound' extends keyof T ? T['rejectOnNotFound'] : false
 > {
   /**
    * @private
    */
   private fetcher;
   /**
    * @private
    */
   private readonly dmmf;
   /**
    * @private
    */
   private connectionPromise?;
   /**
    * @private
    */
   private disconnectionPromise?;
   /**
    * @private
    */
   private readonly engineConfig;
   /**
    * @private
    */
   private readonly measurePerformance;
 
   /**
    * ##  Prisma Client ʲˢ
    *
    * Type-safe database client for TypeScript & Node.js
    * @example
    * ```
    * const prisma = new PrismaClient()
    * // Fetch zero or more Users
    * const users = await prisma.user.findMany()
    * ```
    *
    *
    * Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client).
    */
 
   constructor(optionsArg?: Prisma.Subset<T, Prisma.PrismaClientOptions>);
   $on<V extends U | 'beforeExit'>(
     eventType: V,
     callback: (
       event: V extends 'query'
         ? Prisma.QueryEvent
         : V extends 'beforeExit'
         ? () => Promise<void>
         : Prisma.LogEvent
     ) => void
   ): void;
 
   /**
    * Connect with the database
    */
   $connect(): Promise<void>;
 
   /**
    * Disconnect from the database
    */
   $disconnect(): Promise<any>;
 
   /**
    * Add a middleware
    */
   $use(cb: Prisma.Middleware): void;
 
   /**
    * Executes a prepared raw query and returns the number of affected rows.
    * @example
    * ```
    * const result = await prisma.$executeRaw`UPDATE User SET cool = ${true} WHERE email = ${'user@email.com'};`
    * ```
    *
    * Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client/raw-database-access).
    */
   $executeRaw<T = unknown>(
     query: TemplateStringsArray | Prisma.Sql,
     ...values: any[]
   ): PrismaPromise<number>;
 
   /**
    * Executes a raw query and returns the number of affected rows.
    * Susceptible to SQL injections, see documentation.
    * @example
    * ```
    * const result = await prisma.$executeRawUnsafe('UPDATE User SET cool = $1 WHERE email = $2 ;', true, 'user@email.com')
    * ```
    *
    * Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client/raw-database-access).
    */
   $executeRawUnsafe<T = unknown>(query: string, ...values: any[]): PrismaPromise<number>;
 
   /**
    * Performs a prepared raw query and returns the `SELECT` data.
    * @example
    * ```
    * const result = await prisma.$queryRaw`SELECT * FROM User WHERE id = ${1} OR email = ${'user@email.com'};`
    * ```
    *
    * Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client/raw-database-access).
    */
   $queryRaw<T = unknown>(
     query: TemplateStringsArray | Prisma.Sql,
     ...values: any[]
   ): PrismaPromise<T>;
 
   /**
    * Performs a raw query and returns the `SELECT` data.
    * Susceptible to SQL injections, see documentation.
    * @example
    * ```
    * const result = await prisma.$queryRawUnsafe('SELECT * FROM User WHERE id = $1 OR email = $2;', 1, 'user@email.com')
    * ```
    *
    * Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client/raw-database-access).
    */
   $queryRawUnsafe<T = unknown>(query: string, ...values: any[]): PrismaPromise<T>;
 
   /**
    * Allows the running of a sequence of read/write operations that are guaranteed to either succeed or fail as a whole.
    * @example
    * ```
    * const [george, bob, alice] = await prisma.$transaction([
    *   prisma.user.create({ data: { name: 'George' } }),
    *   prisma.user.create({ data: { name: 'Bob' } }),
    *   prisma.user.create({ data: { name: 'Alice' } }),
    * ])
    * ```
    *
    * Read more in our [docs](https://www.prisma.io/docs/concepts/components/prisma-client/transactions).
    */
   $transaction<P extends PrismaPromise<any>[]>(arg: [...P]): Promise<UnwrapTuple<P>>;
 
   /**
    * `prisma.user`: Exposes CRUD operations for the **User** model.
    * Example usage:
    * ```ts
    * // Fetch zero or more Users
    * const users = await prisma.user.findMany()
    * ```
    */
   get user(): Prisma.UserDelegate<GlobalReject>;
 }
 
 export namespace Prisma {
// ...
   type SelectAndInclude = {
     select: any;
     include: any;
   };
   type HasSelect = {
     select: any;
   };
   type HasInclude = {
     include: any;
   };
   type CheckSelect<T, S, U> = T extends SelectAndInclude
     ? 'Please either choose `select` or `include`'
     : T extends HasSelect
     ? U
     : T extends HasInclude
     ? U
     : S;
// ...
   export type Enumerable<T> = T | Array<T>;
 
   export type RequiredKeys<T> = {
     [K in keyof T]-?: {} extends Prisma__Pick<T, K> ? never : K;
   }[keyof T];
 
   export type TruthyKeys<T> = {
     [key in keyof T]: T[key] extends false | undefined | null ? never : key;
   }[keyof T];
 
   export type TrueKeys<T> = TruthyKeys<Prisma__Pick<T, RequiredKeys<T>>>;
 
   /**
    * Subset
    * @desc From `T` pick properties that exist in `U`. Simple version of Intersection
    */
   export type Subset<T, U> = {
     [key in keyof T]: key extends keyof U ? T[key] : never;
   };
 
   /**
    * SelectSubset
    * @desc From `T` pick properties that exist in `U`. Simple version of Intersection.
    * Additionally, it validates, if both select and include are present. If the case, it errors.
    */
   export type SelectSubset<T, U> = {
     [key in keyof T]: key extends keyof U ? T[key] : never;
   } & (T extends SelectAndInclude ? 'Please either choose `select` or `include`.' : {});
// ...
   type XOR<T, U> = T | U extends object ? (Without<T, U> & U) | (Without<U, T> & T) : T | U;
// ...
   export const type: unique symbol;
// ...
   export type Datasources = {
     db?: Datasource;
   };
 
   export type RejectOnNotFound = boolean | ((error: Error) => Error);
   export type RejectPerModel = { [P in ModelName]?: RejectOnNotFound };
   export type RejectPerOperation = {
     [P in 'findUnique' | 'findFirst']?: RejectPerModel | RejectOnNotFound;
   };
// ...
   export type ErrorFormat = 'pretty' | 'colorless' | 'minimal';
 
   export interface PrismaClientOptions {
     /**
      * Configure findUnique/findFirst to throw an error if the query returns null.
      *  * @example
      * ```
      * // Reject on both findUnique/findFirst
      * rejectOnNotFound: true
      * // Reject only on findFirst with a custom error
      * rejectOnNotFound: { findFirst: (err) => new Error("Custom Error")}
      * // Reject on user.findUnique with a custom error
      * rejectOnNotFound: { findUnique: {User: (err) => new Error("User not found")}}
      * ```
      */
     rejectOnNotFound?: RejectOnNotFound | RejectPerOperation;
     /**
      * Overwrites the datasource url from your prisma.schema file
      */
     datasources?: Datasources;
 
     /**
      * @default "colorless"
      */
     errorFormat?: ErrorFormat;
 
     /**
      * @example
      * ```
      * // Defaults to stdout
      * log: ['query', 'info', 'warn', 'error']
      *
      * // Emit as events
      * log: [
      *  { emit: 'stdout', level: 'query' },
      *  { emit: 'stdout', level: 'info' },
      *  { emit: 'stdout', level: 'warn' }
      *  { emit: 'stdout', level: 'error' }
      * ]
      * ```
      * Read more in our [docs](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-client/logging#the-log-option).
      */
     log?: Array<LogLevel | LogDefinition>;
   }
 
   export type Hooks = {
     beforeRequest?: (options: {
       query: string;
       path: string[];
       rootField?: string;
       typeName?: string;
       document: any;
     }) => any;
   };
 
   /* Types for Logging */
   export type LogLevel = 'info' | 'query' | 'warn' | 'error';
   export type LogDefinition = {
     level: LogLevel;
     emit: 'stdout' | 'event';
   };
 
   export type GetLogType<T extends LogLevel | LogDefinition> = T extends LogDefinition
     ? T['emit'] extends 'event'
       ? T['level']
       : never
     : never;
   export type GetEvents<T extends any> = T extends Array<LogLevel | LogDefinition>
     ? GetLogType<T[0]> | GetLogType<T[1]> | GetLogType<T[2]> | GetLogType<T[3]>
     : never;
 
   export type QueryEvent = {
     timestamp: Date;
     query: string;
     params: string;
     duration: number;
     target: string;
   };
 
   export type LogEvent = {
     timestamp: Date;
     message: string;
     target: string;
   };
// ...
   export type Middleware<T = any> = (
     params: MiddlewareParams,
     next: (params: MiddlewareParams) => Promise<T>
   ) => Promise<T>;
// ...
   export type UserSelect = {
     id?: boolean;
     email?: boolean;
     name?: boolean;
     role?: boolean;
   };
 
   export type UserGetPayload<
     S extends boolean | null | undefined | UserArgs,
     U = keyof S
   > = S extends true
     ? User
     : S extends undefined
     ? never
     : S extends UserArgs | UserFindManyArgs
     ? 'include' extends U
       ? User
       : 'select' extends U
       ? {
           [P in TrueKeys<S['select']>]: P extends keyof User ? User[P] : never;
         }
       : User
     : User;
// ...
   export type UserFindFirstArgs = {
     /**
      * Select specific fields to fetch from the User
      *
      **/
     select?: UserSelect | null;
     /**
      * Throw an Error if a User can't be found
      *
      **/
     rejectOnNotFound?: RejectOnNotFound;
     /**
      * Filter, which User to fetch.
      *
      **/
     where?: UserWhereInput;
     /**
      * {@link https://www.prisma.io/docs/concepts/components/prisma-client/sorting Sorting Docs}
      *
      * Determine the order of Users to fetch.
      *
      **/
     orderBy?: Enumerable<UserOrderByInput>;
     /**
      * {@link https://www.prisma.io/docs/concepts/components/prisma-client/pagination#cursor-based-pagination Cursor Docs}
      *
      * Sets the position for searching for Users.
      *
      **/
     cursor?: UserWhereUniqueInput;
     /**
      * {@link https://www.prisma.io/docs/concepts/components/prisma-client/pagination Pagination Docs}
      *
      * Take `±n` Users from the position of the cursor.
      *
      **/
     take?: number;
     /**
      * {@link https://www.prisma.io/docs/concepts/components/prisma-client/pagination Pagination Docs}
      *
      * Skip the first `n` Users.
      *
      **/
     skip?: number;
     /**
      * {@link https://www.prisma.io/docs/concepts/components/prisma-client/distinct Distinct Docs}
      *
      * Filter by unique combinations of Users.
      *
      **/
     distinct?: Enumerable<UserScalarFieldEnum>;
   };
// ...
   export type UserCreateArgs = {
     /**
      * Select specific fields to fetch from the User
      *
      **/
     select?: UserSelect | null;
     /**
      * The data needed to create a User.
      *
      **/
     data: XOR<UserCreateInput, UserUncheckedCreateInput>;
   };
// ...
   export type UserArgs = {
     /**
      * Select specific fields to fetch from the User
      *
      **/
     select?: UserSelect | null;
   };
 
   /**
    * Enums
    */
 
   // Based on
   // https://github.com/microsoft/TypeScript/issues/3192#issuecomment-261720275
 
   export const UserScalarFieldEnum: {
     id: 'id';
     email: 'email';
     name: 'name';
     role: 'role';
   };
 
   export type UserScalarFieldEnum = typeof UserScalarFieldEnum[keyof typeof UserScalarFieldEnum];
 
   export const SortOrder: {
     asc: 'asc';
     desc: 'desc';
   };
 
   export type SortOrder = typeof SortOrder[keyof typeof SortOrder];
 
   /**
    * Deep Input Types
    */
 
   export type UserWhereInput = {
     AND?: Enumerable<UserWhereInput>;
     OR?: Enumerable<UserWhereInput>;
     NOT?: Enumerable<UserWhereInput>;
     id?: IntFilter | number;
     email?: StringFilter | string;
     name?: StringNullableFilter | string | null;
     role?: EnumRolesFilter | Roles;
   };
 
   export type UserOrderByInput = {
     id?: SortOrder;
     email?: SortOrder;
     name?: SortOrder;
     role?: SortOrder;
   };
 
   export type UserWhereUniqueInput = {
     id?: number;
     email?: string;
   };
// ...
   export type UserCreateInput = {
     email: string;
     name?: string | null;
     role: Roles;
   };
 
   export type UserUncheckedCreateInput = {
     id?: number;
     email: string;
     name?: string | null;
     role: Roles;
   };
// ...
   export const dmmf: runtime.DMMF.Document;
 }